assets: add pages settings page
* refactor community forms components
