+++ invenio_communities/assets/semantic-ui/js/invenio_communities/settings/components/index.js
@@ -1,0 +1,1 @@
+export { CommunitySettingsForm } from "./CommunitySettingsForm";
