+++ invenio_communities/assets/semantic-ui/js/invenio_communities/settings/curationPolicy/index.js
