+++ invenio_communities/config.py
@@ -24,6 +24,7 @@ COMMUNITIES_ROUTES = {
     "requests": "/communities/<pid_value>/requests",
     "settings_privileges": "/communities/<pid_value>/settings/privileges",
     "settings_curation_policy": "/communities/<pid_value>/settings/curation-policy",
+    "settings_pages": "/communities/<pid_value>/settings/pages",
     "members": "/communities/<pid_value>/members",
     "invitations": "/communities/<pid_value>/invitations",
     "about": "/communities/<pid_value>/about",
