+++ invenio_communities/templates/semantic-ui/invenio_communities/details/settings/base.html
@@ -18,6 +18,7 @@
       'profile': (_('Profile'), url_for('invenio_communities.communities_settings', pid_value=community.slug), ),
       'privileges': (_('Privileges'), url_for('invenio_communities.communities_settings_privileges', pid_value=community.slug)),
       'curation_policy': (_('Curation policy'), url_for('invenio_communities.communities_settings_curation_policy', pid_value=community.slug)),
+      'pages': (_('Pages'), url_for('invenio_communities.communities_settings_pages', pid_value=community.slug)),
     } %}
     <div class="sixteen wide mobile sixteen wide tablet three wide computer column">
       <div class="ui vertical horizontal mobile tablet menu">
