+++ invenio_communities/views/communities.py
@@ -252,6 +252,23 @@ def communities_settings_curation_policy(pid_value, community, community_ui):
 
 
 @pass_community(serialize=True)
+def communities_settings_pages(pid_value, community, community_ui):
+    """Community settings/curation-policy page."""
+    permissions = community.has_permissions_to(["update", "read", "search_requests", "search_invites"])
+    if not permissions["can_update"]:
+        raise PermissionDeniedError()
+
+    return render_template(
+        "invenio_communities/details/settings/pages.html",
+        community=community_ui,
+        permissions=permissions,
+        form_config=dict(
+            access=dict(review_policy=REVIEW_POLICY_FIELDS),
+        ),
+    )
+
+
+@pass_community(serialize=True)
 def members(pid_value, community, community_ui):
     """Community members page."""
     permissions = community.has_permissions_to(
