+++ invenio_communities/webpack.py
@@ -27,9 +27,10 @@ communities = WebpackThemeBundle(
         "semantic-ui": dict(
             entry={
                 "invenio-communities-new": "./js/invenio_communities/community/new.js",
-                "invenio-communities-privileges": "./js/invenio_communities/settings/privileges.js",
+                "invenio-communities-privileges": "./js/invenio_communities/settings/priviledges/index.js",
                 "invenio-communities-profile": "./js/invenio_communities/settings/profile/index.js",
-                "invenio-communities-curation-policy": "./js/invenio_communities/settings/curation-policy/index.js",
+                "invenio-communities-curation-policy": "./js/invenio_communities/settings/curationPolicy/index.js",
+                "invenio-communities-settings-pages": "./js/invenio_communities/settings/pages/index.js",
                 "invenio-communities-requests": "./js/invenio_communities/requests/index.js",
                 "invenio-communities-frontpage": "./js/invenio_communities/community/frontpage.js",
                 "invenio-communities-search": "./js/invenio_communities/community/search.js",
